+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - [+] #20 allow jinja filters and globals to be extended using `jinja_helpers`
   module
 - [+] #18 page and post configuration can have custom template for rendering
-- [+] #4, #16 local links validation
+- [+] #4, #16, #22 local links validation
 - [+] #7 add option to skip feed rendering
 - [\*] #8 render mainpage only once in case mainpage content is in pages
 - [+] #2 pagination for posts (mainpage) list